backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==23.2
action==1.4.4
django-import-export==3.3.1
python-dotenv
django-colorfield==0.10.1
djoser==2.2.0
drf-extra-fields==3.7.0
reportlab==4.0.5
load_dotenv==0.1.0